pyproject.toml
[project]
name = "ideasforindians"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "firebase-admin==6.3.0",
    "flask==3.0.0",
    "gunicorn==21.2.0",
    "python-dotenv==1.0.0",
    "requests==2.31.0",
    "werkzeug==3.0.1",
]
